+++ src/job_aggregator/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Plugin subpackage for job-aggregator source plugins."""